color_detector/requirements.txt
tkinter
colorthief
pillow